refactor(processing): extract job display and input staging from JobsModule
JobsModule.vue had accumulated three jobs beyond orchestration: pure
display-string formatting, store-reading input staging, and the
binding/validation glue that is actually the component's role. Split
the first two out:

- engine/jobDisplay.ts: pure formatting for submitted-job display and
  bound source-ref names. Group names are resolved into the context by
  the caller, and the previously duplicated source-ref dispatch in
  boundSourceRefName/formatProcessingValue is unified with per-type
  fallbacks.
- composables/useInputStaging.ts: active-image input reads and the
  Run-time staging of segment groups and annotations, including the
  annotation tools view. stageLabelmapInputs now requires its bindings
  instead of defaulting to a fresh bind.

No behavior change.
