+++ src/processing/components/JobsModule.vue
@@ -126,62 +126,31 @@ import { useCropStore } from '@/src/store/tools/crop';
 import type {
   ProcessingProvider,
   ProcessingValue,
-  SubmittedJobDisplay,
-  SubmittedJobParameterDisplay,
   TaskSummary,
 } from '@/src/processing/types';
 import {
   buildTaskFormModel,
   initialFormValues,
   validateFormValues,
-  fieldLabel,
   type TaskFormModel,
   type FormValidationIssue,
 } from '@/src/processing/engine/formModel';
 import { type SourceRefBindingState } from '@/src/processing/engine/mintInput';
 import {
-  mintLabelmapValue,
-  mintLabelmapReferenceImage,
-  type SegmentGroupView,
-} from '@/src/processing/engine/mintLabelmap';
-import { mintInputValue } from '@/src/processing/engine/mintInput';
-import { mintAnnotationsValue } from '@/src/processing/engine/mintAnnotations';
-import {
-  annotationToolsViewCount,
-  annotationsFileCount,
-  encodeAnnotationsFile,
-  hasTwoPoints,
-  isEncodablePolygon,
-  type AnnotationKindView,
-  type AnnotationToolsView,
-  type PolygonToolView,
-  type TwoPointToolView,
-} from '@/src/processing/engine/annotationsWire';
-import { annotationToolStore } from '@/src/processing/annotationKinds';
-import {
   bindSourceRefs,
   type BoundSourceRefType,
   type SourceRefBindings,
 } from '@/src/processing/engine/sourceRefs';
+import {
+  buildJobDisplay,
+  buildSourceRefNames,
+  type JobDisplayContext,
+} from '@/src/processing/engine/jobDisplay';
 import { cropPlanesToWorldBounds } from '@/src/processing/engine/bounds';
+import { useInputStaging } from '@/src/processing/composables/useInputStaging';
 import { usePaintToolStore } from '@/src/store/tools/paint';
 import { useSegmentGroupStore } from '@/src/store/segmentGroups';
 import { useMessageStore } from '@/src/store/messages';
-import { writeSegmentation } from '@/src/io/readWriteImage';
-import { getDataSourceName } from '@/src/io/import/dataSource';
-import type {
-  AnnotationToolKind,
-  AnnotationsFile,
-  InputValue,
-  VolViewTaskParameter,
-} from '@/backend-contract';
-import {
-  ANNOTATIONS_FILE_EXTENSION,
-  TYPE_TAG_ANNOTATIONS,
-  TYPE_TAG_LABELMAP,
-} from '@/backend-contract';
-import type { AnnotationTool } from '@/src/types/annotation-tool';
-import { stripExtension } from '@/src/utils/path';
 
 import TaskPicker from './TaskPicker.vue';
 import TaskForm from './TaskForm.vue';
@@ -195,6 +164,16 @@ const cropStore = useCropStore();
 const paintStore = usePaintToolStore();
 const segmentGroupStore = useSegmentGroupStore();
 const messageStore = useMessageStore();
+
+const {
+  activeDataSource,
+  activeImageName,
+  segmentGroupView,
+  finishedAnnotationCount,
+  captureAnnotationsPayload,
+  stageLabelmapInputs,
+  stageAnnotationInputs,
+} = useInputStaging();
 
 const openPanels = ref<string[]>(['run', 'jobs']);
 
@@ -373,7 +352,11 @@ async function onSubmit(values: Record<string, ProcessingValue>) {
   }
   // Display formatting reads live active image and segment-group state, so it
   // must render before the staging await.
-  const display = buildJobDisplay(model, finalValues);
+  const display = buildJobDisplay(
+    model,
+    jobDisplayContext(bindings),
+    finalValues
+  );
   // Same reason, and it also outlives the staging awaits below: the annotations
   // file, its name, and its reference image all come from the active image,
   // which the user may switch while an upload is in flight. Only encoded when
@@ -405,7 +388,7 @@ async function onSubmit(values: Record<string, ProcessingValue>) {
   try {
     staged = await Promise.all([
       stage('Failed to stage segment group input', () =>
-        stageLabelmapInputs(submitProvider, model, bindings)
+        stageLabelmapInputs(submitProvider, bindings)
       ),
       stage('Failed to stage annotations input', () =>
         stageAnnotationInputs(submitProvider, bindings, annotations)
@@ -458,69 +441,6 @@ function applyBoundsBindings(
   });
   return next;
 }
-
-function activeDataSource() {
-  return datasetStore.getDataSource(currentImageID.value);
-}
-
-function activeImageName(): string | undefined {
-  const id = currentImageID.value;
-  return (
-    imageCache.getImageMetadata(id)?.name ??
-    getDataSourceName(activeDataSource()) ??
-    undefined
-  );
-}
-
-function segmentGroupView(): SegmentGroupView {
-  return {
-    orderByParent: segmentGroupStore.orderByParent,
-    metadataByID: segmentGroupStore.metadataByID,
-  };
-}
-
-function labelmapReferenceImage(segmentGroupId: string): InputValue | null {
-  return mintLabelmapReferenceImage(
-    segmentGroupId,
-    segmentGroupView(),
-    (imageId) => datasetStore.getDataSource(imageId)
-  );
-}
-
-// Tool lists are per image, and so is the staged annotations file: only the
-// active image's finished tools are ever an input.
-function onActiveImage<T extends { imageID: string }>(
-  tools: readonly T[]
-): T[] {
-  const id = currentImageID.value;
-  return id ? tools.filter((tool) => tool.imageID === id) : [];
-}
-
-// The three stores are independent, so each keeps its own label namespace; the
-// encoder prunes and re-keys them by name.
-const annotationToolsView = computed<AnnotationToolsView>(() => {
-  const kindView = <T extends object>(
-    kind: AnnotationToolKind,
-    hasGeometry: <U extends AnnotationTool>(tool: U) => tool is U & T
-  ): AnnotationKindView<AnnotationTool & T> => {
-    const store = annotationToolStore(kind);
-    return {
-      tools: onActiveImage(store.finishedTools).filter(hasGeometry),
-      labels: store.labels,
-    };
-  };
-  return {
-    rulers: kindView<TwoPointToolView>('rulers', hasTwoPoints),
-    rectangles: kindView<TwoPointToolView>('rectangles', hasTwoPoints),
-    polygons: kindView<PolygonToolView>('polygons', isEncodablePolygon),
-  };
-});
-
-// Computed, not a function call: placing a tool churns the stores every drag
-// frame, and an unchanged count stops the invalidation there.
-const finishedAnnotationCount = computed(() =>
-  annotationToolsViewCount(annotationToolsView.value)
-);
 
 function activeSourceBindings(model: TaskFormModel): SourceRefBindings {
   return bindSourceRefs(model, {
@@ -582,255 +502,35 @@ function refreshValidation(
   return validation.issues;
 }
 
-// Returns only the parameters it staged, so the caller owns the merge.
-//
-// The literal 'seg.nrrd' name is required for segment names and colors to be
-// embedded in the serialized output.
-async function stageSegmentGroupInput(
-  p: ProcessingProvider,
-  segmentGroupId: string,
-  // Group names are not unique, so a plural parameter numbers its files to keep
-  // them distinguishable in the job folder.
-  ordinal?: number
-): Promise<string[]> {
-  const metadata = segmentGroupStore.metadataByID[segmentGroupId];
-  const labelmap = segmentGroupStore.dataIndex[segmentGroupId];
-  const referenceImage = labelmapReferenceImage(segmentGroupId);
-  if (!referenceImage) {
-    throw new Error('Segment group reference image has no server provenance');
-  }
-  const serialized = await writeSegmentation('seg.nrrd', labelmap, metadata);
-  const suffix = ordinal === undefined ? '' : `-${ordinal}`;
-  return p.stageInput({
-    file: new Blob([serialized]),
-    descriptor: {
-      type: TYPE_TAG_LABELMAP,
-      name: `${metadata.name}${suffix}.seg.nrrd`,
-      referenceImage: {
-        ...referenceImage,
-        type: 'image',
-      },
-    },
-  });
-}
-
-async function stageLabelmapInputs(
-  p: ProcessingProvider,
-  model: TaskFormModel,
-  bindings: SourceRefBindings = activeSourceBindings(model)
-): Promise<Record<string, ProcessingValue>> {
-  const targets = Object.entries(bindings.labelmap.groups);
-  if (targets.length === 0) return {};
-
-  const staged = await Promise.all(
-    targets.map(async ([parameterId, segmentGroupIds]) => {
-      const uris = (
-        await Promise.all(
-          segmentGroupIds.map((groupId, index) =>
-            stageSegmentGroupInput(
-              p,
-              groupId,
-              segmentGroupIds.length > 1 ? index + 1 : undefined
-            )
-          )
-        )
-      ).flat();
-      return [parameterId, mintLabelmapValue(uris)] as const;
-    })
-  );
-  return Object.fromEntries(staged);
-}
-
-// The extension is what the CLI spec and the backend both match on, so the base
-// name is the active image's without its own — compound extensions included, so
-// `scan.nii.gz` stages as `scan.annotations.json`.
-function annotationsFileName(): string {
-  const name = activeImageName() ?? 'image';
-  return `${stripExtension(name)}${ANNOTATIONS_FILE_EXTENSION}`;
-}
-
-// Everything the annotations file is made of, read off the stores in one
-// synchronous pass so staging never mixes two images' state.
-type AnnotationsPayload = {
-  file: AnnotationsFile;
-  name: string;
-  referenceImage: InputValue | null;
-};
-
-function captureAnnotationsPayload(): AnnotationsPayload {
-  return {
-    file: encodeAnnotationsFile(annotationToolsView.value),
-    name: annotationsFileName(),
-    referenceImage: mintInputValue(activeDataSource()),
-  };
-}
-
-// One file per bound parameter, holding every finished tool the active image
-// had at Run. The image is its own reference image, so a volume without server
-// provenance never gets here — the binder already refused it.
-async function stageAnnotationInputs(
-  p: ProcessingProvider,
-  bindings: SourceRefBindings,
-  payload: AnnotationsPayload | null
-): Promise<Record<string, ProcessingValue>> {
-  const [parameterId] = bindings.annotations.parameters;
-  if (!parameterId || !payload) return {};
-  const { file, name, referenceImage } = payload;
-
-  if (!referenceImage) {
-    throw new Error('The active image has no server provenance');
-  }
-  // The binder validated a live count; this is the encoded file's own count, so
-  // a tool deleted between validation and Run cannot stage an empty file.
-  if (annotationsFileCount(file) === 0) {
-    throw new Error('The active image has no finished annotations');
-  }
-
-  const uris = await p.stageInput({
-    file: new Blob([JSON.stringify(file)], { type: 'application/json' }),
-    descriptor: {
-      type: TYPE_TAG_ANNOTATIONS,
-      name,
-      referenceImage: {
-        ...referenceImage,
-        type: 'image',
-      },
-    },
-  });
-  return {
-    [parameterId]: mintAnnotationsValue(uris),
-  };
-}
-
-type SourceRefContext = {
-  labelmapGroups: Record<string, string[]>;
-  types: Record<string, BoundSourceRefType>;
-  imageName: string | undefined;
-  annotationCount: number;
-};
-
 // Resolved once per display pass: each binding re-runs a full field scan and
 // group resolution, so per-field resolution would redo identical work.
-function sourceRefContext(model: TaskFormModel): SourceRefContext {
-  const bindings = activeSourceBindings(model);
+function jobDisplayContext(bindings: SourceRefBindings): JobDisplayContext {
+  const labelmapNames = Object.fromEntries(
+    Object.entries(bindings.labelmap.groups).map(([parameterId, groupIds]) => [
+      parameterId,
+      groupIds.map(
+        (groupId) =>
+          segmentGroupStore.metadataByID[groupId]?.name ??
+          'unnamed segment group'
+      ),
+    ])
+  );
   return {
-    labelmapGroups: bindings.labelmap.groups,
+    labelmapNames,
     types: bindings.types,
     imageName: activeImageName(),
     annotationCount: finishedAnnotationCount.value,
   };
 }
 
-function boundLabelmapName(
-  refs: SourceRefContext,
-  parameterId: string
-): string | undefined {
-  const groupIds = refs.labelmapGroups[parameterId] ?? [];
-  const names = groupIds.map(
-    (groupId) =>
-      segmentGroupStore.metadataByID[groupId]?.name ?? 'unnamed segment group'
-  );
-  return names.length > 0 ? names.join(', ') : undefined;
-}
-
-// The bound value is a whole set of tools rather than one named resource, so
-// the count is the identifying part.
-function boundAnnotationsName(refs: SourceRefContext): string {
-  const noun = refs.annotationCount === 1 ? 'annotation' : 'annotations';
-  const count = `${refs.annotationCount} ${noun}`;
-  return refs.imageName ? `${count} on ${refs.imageName}` : count;
-}
-
-function boundSourceRefName(
-  refs: SourceRefContext,
-  parameterId: string
-): string | undefined {
-  const type = refs.types[parameterId];
-  if (type === TYPE_TAG_LABELMAP) return boundLabelmapName(refs, parameterId);
-  if (type === TYPE_TAG_ANNOTATIONS) return boundAnnotationsName(refs);
-  return refs.imageName;
-}
-
 const sourceRefNames = computed(() => {
   const model = taskModel.value;
   if (!model) return {};
-  const refs = sourceRefContext(model);
-  const names: Record<string, string> = {};
-  model.fields.forEach((field) => {
-    if (field.kind !== 'sourceRef') return;
-    const name = boundSourceRefName(refs, field.id);
-    if (name) names[field.id] = name;
-  });
-  return names;
-});
-
-function formatProcessingValue(
-  refs: SourceRefContext,
-  field: VolViewTaskParameter,
-  value: ProcessingValue
-): string {
-  if (field.kind === 'sourceRef') {
-    if (refs.types[field.id] === TYPE_TAG_LABELMAP) {
-      // A bound param always names its groups, so the fallback is the optional
-      // param that bound nothing.
-      return boundLabelmapName(refs, field.id) ?? 'not provided';
-    }
-    if (refs.types[field.id] === TYPE_TAG_ANNOTATIONS) {
-      return boundAnnotationsName(refs);
-    }
-    return refs.imageName ?? 'active dataset';
-  }
-  if (field.kind === 'bounds') {
-    return Array.isArray(value) && value.length > 0
-      ? value.map((n) => (typeof n === 'number' ? n.toFixed(1) : n)).join(', ')
-      : 'not set';
-  }
-  if (typeof value === 'boolean') return value ? 'true' : 'false';
-  if (Array.isArray(value)) return value.join(', ');
-  if (value && typeof value === 'object') {
-    const input = value as InputValue;
-    return input.type;
-  }
-  if (value === null || value === undefined || value === '') return 'not set';
-  return String(value);
-}
-
-function isSummaryParameter(
-  field: VolViewTaskParameter,
-  value: ProcessingValue
-): boolean {
-  if (field.kind === 'sourceRef' || field.kind === 'bounds') return false;
-  if (value === null || value === undefined || value === '') return false;
-  if (Array.isArray(value) && value.length === 0) return false;
-  return true;
-}
-
-function buildJobDisplay(
-  model: TaskFormModel,
-  values: Record<string, ProcessingValue>
-): SubmittedJobDisplay {
-  const refs = sourceRefContext(model);
-  let summaryCount = 0;
-  const parameters: SubmittedJobParameterDisplay[] = model.fields.map(
-    (field) => {
-      const value = values[field.id];
-      const summary = summaryCount < 2 && isSummaryParameter(field, value);
-      if (summary) summaryCount += 1;
-      return {
-        id: field.id,
-        label: fieldLabel(field),
-        value: formatProcessingValue(refs, field, value),
-        ...(summary ? { summary } : {}),
-      };
-    }
+  return buildSourceRefNames(
+    model,
+    jobDisplayContext(activeSourceBindings(model))
   );
-  const inputName = refs.imageName;
-  return {
-    taskTitle: model.title,
-    ...(inputName ? { inputName } : {}),
-    parameters,
-  };
-}
+});
 
 // Debounced: dragging a crop handle mutates the crop planes every frame, and
 // each rebind walks provenance and re-validates the whole form.
